Логотип вынесен в отдельный компонент
