+++ src/components/header/header.jsx
@@ -1,24 +1,14 @@
 import PropTypes from 'prop-types';
 import React from "react";
 import {Link} from 'react-router-dom';
+import Logo from '../logo/logo';
 
 const Header = (props) =>{
-  const {isLinkActive, className, showUserBlock} = props;
-
-  const logo = (
-    <React.Fragment>
-      <span className="logo__letter logo__letter--1">W</span>
-      <span className="logo__letter logo__letter--2">T</span>
-      <span className="logo__letter logo__letter--3">W</span>
-    </React.Fragment>);
+  const {className, showUserBlock} = props;
 
   return (
     <header className={`page-header ${className}`}>
-      <div className="logo">
-        {isLinkActive
-          ? <Link className="logo__link" to="/">{logo}</Link>
-          : <a className="logo__link">{logo}</a>}
-      </div>
+      <Logo/>
       {props.children}
       <div className="user-block">
         {showUserBlock &&
@@ -36,14 +26,12 @@ const Header = (props) =>{
 
 Header.defaultProps = {
   classNames: ``,
-  isLinkActive: true,
   showUserBlock: true
 };
 
 Header.propTypes = {
   children: PropTypes.element,
-  className: PropTypes.arrayOf(PropTypes.string),
-  isLinkActive: PropTypes.bool,
+  className: PropTypes.string,
   showUserBlock: PropTypes.bool,
 };
